+++ packages/components/src/remix-hook-form/index.ts
@@ -8,3 +8,7 @@ export * from './radio-group';
 export * from './radio-group-item';
 export * from './switch';
 export * from './textarea';
+export * from './middleware';
+export * from './data-table-router-form';
+export * from './data-table-router-parsers';
+export * from './data-table-router-toolbar';